Linaro string handling routines: Give a option to use the older method
of memcpy for device compatibility testing.  The newer method of memcpy
provides a nice boost in performance.
Use export USE_OLD_MEMCPY=true
to use the older method of memcpy if you run into any issues.
For ARM deivces that are not ARMV7A the newer method of memcpy will
be used either way.

Signed-off-by:  Paul Beeler <sparksco@gmail.com>

Include bzero.S in all variants for ARMV7A/NON_NEON_MEMCPY devices.
This fixes the game plague on the newer method of memcpy which was
previously broken.  Tested on grouper.  Might still be broken on
devices that use neon memcpy.

TODO *//More Testing

Reported-by: @xda user matthewdavis
Signed-off-by:  Paul Beeler <sparksco@gmail.com>